[Chore] Move `MediaConnector` to `vllm.multimodal.media`
