+++ vllm/multimodal/media/__init__.py
@@ -2,8 +2,9 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 from .audio import AudioEmbeddingMediaIO, AudioMediaIO
 from .base import MediaIO, MediaWithBytes
+from .connector import MEDIA_CONNECTOR_REGISTRY, MediaConnector
 from .image import ImageEmbeddingMediaIO, ImageMediaIO
-from .video import VideoMediaIO
+from .video import VIDEO_LOADER_REGISTRY, VideoMediaIO
 
 __all__ = [
     "MediaIO",
@@ -12,5 +13,8 @@ __all__ = [
     "AudioMediaIO",
     "ImageEmbeddingMediaIO",
     "ImageMediaIO",
+    "VIDEO_LOADER_REGISTRY",
     "VideoMediaIO",
+    "MEDIA_CONNECTOR_REGISTRY",
+    "MediaConnector",
 ]
